Rename bftcosi->byzcoinx, cosi->ftcosi
