+++ ftcosi/ftcosi_test.go
