+++ ftcosi/protocol/helper_functions.go
