+++ ftcosi/protocol/protocol_test.go
