+++ ftcosi/server.go
@@ -6,8 +6,8 @@ import (
 	// Empty imports to have the init-functions called which should
 	// register the protocol
 
-	_ "github.com/dedis/cothority/cosi/protocol"
-	_ "github.com/dedis/cothority/cosi/service"
+	_ "github.com/dedis/cothority/ftcosi/protocol"
+	_ "github.com/dedis/cothority/ftcosi/service"
 	"github.com/dedis/onet/app"
 )
