+++ ftcosi/simulation/fd_test.go
