+++ ftcosi/simulation/simul.go
