+++ pop/app.go
@@ -12,9 +12,9 @@ import (
 	"strings"
 
 	"github.com/dedis/cothority"
-	"github.com/dedis/cothority/cosi/check"
-	_ "github.com/dedis/cothority/cosi/protocol"
-	_ "github.com/dedis/cothority/cosi/service"
+	"github.com/dedis/cothority/ftcosi/check"
+	_ "github.com/dedis/cothority/ftcosi/protocol"
+	_ "github.com/dedis/cothority/ftcosi/service"
 
 	"github.com/BurntSushi/toml"
 	"github.com/dedis/cothority/pop/service"
