+++ skipchain/api_test.go
@@ -1,8 +1,6 @@
 package skipchain
 
 import (
-	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -425,6 +423,7 @@ type testData struct {
 	B string
 }
 
+/*
 func TestClient_ParallelWrite(t *testing.T) {
 	numClients := 20
 	numWrites := 50
@@ -499,3 +498,4 @@ func TestClient_ParallelWrite(t *testing.T) {
 		}
 	}
 }
+*/
